% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examine.R
\name{examine}
\alias{examine}
\title{Examine the design}
\usage{
examine(effect = 1, se = 1, df = 1, n = 100)
}
\arguments{
\item{effect}{assumed effect size. Numeric value.}

\item{se}{assume standard error. Positive, numeric value.}

\item{df}{degrees of freedom.}

\item{n}{Number of simulated trials.}
}
\value{
statistics
}
\description{
Examine the design
}
\examples{
examine(1, 1, 1, 1)
}
